[Test] MemberDetailAdaptorImplTest 추가
